atlasgpt-backend/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
openai==1.78.1
pydantic==2.11.4
python-dotenv==1.1.0
requests==2.32.3
